requirements.txt
awscurl==0.29
boto3==1.26.131
botocore==1.29.131
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
ConfigArgParse==1.5.3
configparser==5.3.0
cryptography==40.0.2
elasticsearch==6.8.2
idna==3.4
jmespath==1.0.1
pycparser==2.21
pyOpenSSL==23.1.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
requests-aws4auth==1.1.2
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.12
urllib3-secure-extra==0.1.0
python-dotenv